re-organize pkg loader; needs proper cleanup
